+++ flipt-client-node/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@flipt-io/flipt-client",
-  "version": "0.11.0",
+  "version": "0.11.1",
   "description": "Flipt Client Evaluation SDK for Node.js",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
